Bugfix: Fatal error: Cannot use object of type PDOStatement ...
This change fixes a bug which results in the following error message:
PHP Fatal error: Cannot use object of type PDOStatement as array
in tagsets/email.php on line 110

The reason is the dual use of the variable $result for two purposes in the same function.
The fix changes the variable name used for one purpose.