Stylefix: Replace tabs with spaces.
This fix changes the coding style and not affecting the functionality or
changing the order of code lines in any way. It has been agreed that the code
will use 4 spaces indent instead of single tab. This commit implements this
change. This is needed to make code cleaner and easier to work with, especially
when reviewing contributions. From now on all authors contributing the code
containing tabs will be kindly asked to amend and resubmit their pull requests.
The documentation on coding style will be published in project wiki.

Commands used:
    find . -name '*.php' ! -type d -exec bash -c 'expand -t 4 "$0" > /tmp/e && mv /tmp/e "$0"' {} \;
    find ./install/ ! -type d ! -name '*.sql' -exec bash -c 'expand -t 4 "$0" > /tmp/e && mv /tmp/e "$0"' {} \;