Stylefix: Remove trailing spaces.
This fix changes the coding style and not affecting the functionality or
changing the order of code lines in any way. The change removes all trailing
spaces. This is needed to make code cleaner and easier to work with, especially
when reviewing contributions. From now on all authors contributing the code
containing trailing spaces will be kindly asked to amend and resubmit their
pull requests. The documentation on coding style will be published in project
wiki.

Commands used:
    find . -name '*.php' ! -type d -print0 | xargs -0 sed -i 's/[ \t]*$//'
    find ./install/ ! -type d ! -name '*.sql' -print0 | xargs -0 sed -i 's/[ \t]*$//'