Bugfix: Rules signed checkbox not displayed in participant list
Although the “rules checkbox” is enabled in Options/General Settings/Extra features, there is no checkbox for rules displayed in the participant list for a session. Rather, just a text (yes/no) appears in the column. As a result, when the participant list is saved, the status of “rules_signed” is set to “n” for all participants in the list.

This fix adds the checkbox to each row of the participant list and thereby fixes both the display as well as the data saving issue.