Feature: Allow restrict no-show calc to sessions after a particular date Upon user request, this feature allows to restrict the count of no-shows to experiment sessions which took place after a particular date. To do this, we added two new options to General Settings: Enable restriction or not, and set the respective date from which on no-shows will be counted. The cronjob that calculates and updates participants' history was amended such that it if the restriction is enabled and a date has been set, it only considers sessions after that date when calculating the number of no-shows.
I note that the participant history (number of previous registrations, number of previous no-shows) is regularly calculated and updated by a cron job based on all previous participation data. That implies that a change of the setting described above will take effect with the next run of the cron job, and apply to all participations in the database.

On the way, we had to add a new option type "date" to the collection of possible option types. We also had to add a fix to the datepicker since an array variable as a form element name would break the date field code. Future code that uses the datepicker can now use an array element as the name for the form field, but will have to piece the submitted values together afterwards (see changes to top of admin/options_edit.php).